Moving SyncClient functionality into an Abstract class.
